services/ws_frontend/pyproject.toml
[tool.poetry]
name = "ws_frontend"
version = "0.1.0"
description = "Websockets frontend for game process"
authors = ["chiselko6 <alex.limontov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
python-jose = "^3.2.0"
pydantic = "^1.8.1"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}


[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
mypy = "^0.812"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
